+++ efx/src/interfaces/mod.rs
@@ -1,0 +1,7 @@
+pub mod block;
+pub mod tag;
+pub mod tag_attributes;
+
+pub use block::Block;
+pub use tag::Tag;
+pub use tag_attributes::TagAttributes;