+++ efx/src/interfaces/tag_attributes.rs
@@ -1,0 +1,8 @@
+use efx_core::Element;
+use proc_macro2::TokenStream;
+
+pub trait TagAttributes {
+    fn new(el: &Element) -> Result<Self, TokenStream>
+    where
+        Self: Sized;
+}