+++ efx/src/tags/image.rs
@@ -1,4 +1,4 @@
-use crate::tags::{Tag, TagAttributes};
+use crate::interfaces::*;
 use crate::utils::attr::*;
 use crate::utils::expr::*;
 use efx_attrnames::AttrNames;