+++ efx/src/tags/mod.rs
@@ -19,13 +19,10 @@ pub use bottom_panel::BottomPanel;
 pub use button::Button;
 pub use central_panel::CentralPanel;
 pub use column::Column;
-use efx_core::Element;
 pub use heading::Heading;
 pub use hyperlink::Hyperlink;
 pub use image::Image;
 pub use label::Label;
-use proc_macro2::TokenStream;
-use quote::{quote, ToTokens};
 pub use resize::Resize;
 pub use row::Row;
 pub use scroll_area::ScrollArea;
@@ -34,49 +31,3 @@ pub use side_panel::SidePanel;
 pub use text_field::TextField;
 pub use top_panel::TopPanel;
 pub use window::Window;
-
-pub trait Tag: Sized {
-    /// Constructor from Element (parses attributes and captures children inside self).
-    fn from_element(el: &Element) -> Result<Self, TokenStream>
-    where
-        Self: Sized;
-    /// Render contents
-    fn content<UI: ToTokens>(&self, ui: &UI) -> TokenStream;
-    /// Full render
-    fn render<UI: ToTokens>(&self, ui: &UI) -> TokenStream;
-}
-
-pub(crate) trait Block: Sized {
-    fn from_element(el: &Element) -> Result<Self, TokenStream>
-    where
-        Self: Sized;
-    fn content<UI: ToTokens>(&self, ui: &UI) -> TokenStream;
-    fn prolog_epilogue<UI: ToTokens>(&self, ui: &UI) -> (TokenStream, TokenStream);
-
-    /// Full render: prologue → content → epilogue.
-    fn render<UI: ToTokens>(&self, ui: &UI) -> TokenStream {
-        let (prolog, epilogue) = self.prolog_epilogue(ui);
-        let content = self.content(ui);
-        quote! {{ #prolog #content #epilogue }}
-    }
-}
-
-impl<T: Block> Tag for T {
-    fn from_element(el: &Element) -> Result<Self, TokenStream> {
-        <T as Block>::from_element(el)
-    }
-
-    fn content<UI: ToTokens>(&self, ui: &UI) -> TokenStream {
-        <T as Block>::content(self, ui)
-    }
-
-    fn render<UI: ToTokens>(&self, ui: &UI) -> TokenStream {
-        <T as Block>::render(self, ui)
-    }
-}
-
-pub(crate) trait TagAttributes {
-    fn new(el: &Element) -> Result<Self, TokenStream>
-    where
-        Self: Sized;
-}