+++ efx/src/tags/separator.rs
@@ -1,4 +1,4 @@
-use crate::tags::{Block, TagAttributes};
+use crate::interfaces::*;
 use crate::utils::attr::*;
 use efx_attrnames::AttrNames;
 use efx_core::Element;